Change List<Section> to Sections class
