+++ src/main/java/nextstep/subway/applicaion/dto/SectionRequest.java
@@ -17,12 +17,6 @@ public class SectionRequest {
         return distance;
     }
 
-    /*
-        Q. test 위해서 불필요한 Request Constructor가 생성된 것이 아닌가?
-        쓰지 않고 addLine(long, sectionRequest)를 테스트 할 방법이 있는가?
-     */
-
-
     public SectionRequest(Long upStationId, Long downStationId, int distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
